Use Float for money type

+++ app/graphql/types/restaurant_product_type.rb
@@ -3,6 +3,6 @@
 class Types::RestaurantProductType < Types::BaseObject
   field :id, ID, null: false
   field :item, String, null: false
-  field :price, Numeric, null: false
+  field :price, Float, null: false
   field :restaurant, Types::RestaurantType, null: false
 end
